# quizverse/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizVerse API",
        "description": "API dokumentacji dla aplikacji QuizVerse",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/quizzes/{quizId}/questions": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Dodaj pytanie",
                "description": "Dodaje nowe pytanie do quizu (tylko dla właściciela quizu)",
                "operationId": "6ee7bdafc7c83b085197048e652d34a9",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "ID quizu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Jaką jest stolica Polski?"
                                    },
                                    "points": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pytanie zostało dodane",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Question added successfully"
                                        },
                                        "question": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quiz_id": {
                                                    "type": "integer"
                                                },
                                                "text": {
                                                    "type": "string"
                                                },
                                                "points": {
                                                    "type": "integer"
                                                },
                                                "image_path": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/quiz-results": {
            "get": {
                "tags": [
                    "Quiz Results"
                ],
                "summary": "Moje wyniki",
                "description": "Pobiera listę wyników quizów zalogowanego użytkownika",
                "operationId": "c51088cac3dd2f16b5d0fcff73eaca30",
                "responses": {
                    "200": {
                        "description": "Lista wyników",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "quiz_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "points": {
                                                "type": "integer",
                                                "example": 85
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/quiz-results/{id}": {
            "get": {
                "tags": [
                    "Quiz Results"
                ],
                "summary": "Szczegóły wyniku",
                "description": "Pobiera szczegóły wyniku quizu z pytaniami",
                "operationId": "2d7deeb1ac777048f40a8da4b71cf1fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID wyniku",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły wyniku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "points": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wynik nie znaleziony"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}